Fix signed vs unsigned comparison warnings by changing sServerPlayerIndex to a short. Thus synchronizing it with localPlayerIndex and the canonical form of player_index usage in general.
git-svn-id: svn://svn.code.sf.net/p/marathon/code/trunk@1824 a75d40c9-0969-4759-85e5-97cb871b7ed4

+++ Source_Files/Network/network.cpp
@@ -178,7 +178,7 @@ static short localPlayerIndex;
 static short localPlayerIdentifier;
 static NetTopologyPtr topology;
 static char *network_adsp_packet;
-static size_t sServerPlayerIndex;
+static short sServerPlayerIndex;
 static bool sOldSelfSendStatus;
 static RingGameProtocol sRingGameProtocol;
 static StarGameProtocol sStarGameProtocol;
